+++ tests/test_models/test_inference/test_predict.py
@@ -195,6 +195,12 @@ class TestPredictInSampleFull:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_in_sample_full_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -326,6 +332,12 @@ class TestPredictInSampleSuffix:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_in_sample_suffix_datetime_timestamp_failed_not_implemented_predict(
@@ -472,6 +484,12 @@ class TestPredictInSampleSuffix:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_in_sample_suffix_int_timestamp_failed_not_implemented_predict(
@@ -614,6 +632,12 @@ class TestPredictOutSample:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_out_sample_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -773,6 +797,12 @@ class TestPredictOutSamplePrefix:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_out_sample_prefix_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -946,6 +976,12 @@ class TestPredictOutSampleSuffix:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_out_sample_suffix_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -1124,6 +1160,12 @@ class TestPredictMixedInOutSample:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_mixed_in_out_sample_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -1290,6 +1332,12 @@ class TestPredictSubsetSegments:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_subset_segments_failed_not_implemented_predict(self, model, transforms, dataset_name, request):
@@ -1421,6 +1469,12 @@ class TestPredictNewSegments:
             (NBeatsGenericModel(input_size=7, output_size=7, trainer_params=dict(max_epochs=1)), [], "example_tsds"),
             (ChronosModel(path_or_url="amazon/chronos-t5-tiny", encoder_length=7), [], "example_tsds"),
             (ChronosBoltModel(path_or_url="amazon/chronos-bolt-tiny", encoder_length=7), [], "example_tsds"),
+            # pytest.param(
+            #     TimesFMModel(path_or_url="google/timesfm-1.0-200m-pytorch", encoder_length=32),
+            #     [],
+            #     "example_tsds",
+            #     marks=pytest.mark.skip(reason="Model causes OOM in GitHub Actions."),
+            # ),
         ],
     )
     def test_predict_new_segments_failed_not_implemented_predict(self, model, transforms, dataset_name, request):